refactor: gather models
