+++ src/Model/Coin.js
